chore: fix code quality issues found by clippy + SonarQube
- SonarQube: fix async Promise executor, dead code, CSS overflow, sort localeCompare, aria-label, unused assignment
- clippy: auto-fix warnings across engine and tauri crates
- chore: add .scannerwork/ to .gitignore

+++ engine/src/vector/mod.rs
@@ -132,7 +132,7 @@ impl CodeVectorIndex {
         for (slot_key, distance) in results.keys.iter().zip(results.distances.iter()) {
             let slot = *slot_key as usize;
             if slot < slots.len() {
-                let similarity = 1.0 - (*distance as f32).min(2.0).max(0.0);
+                let similarity = 1.0 - (*distance).min(2.0).max(0.0);
                 output.push((slots[slot].clone(), similarity));
             }
         }
